Reorganise fields
