+++ src/api/index.js
@@ -1,89 +1,18 @@
 import arrayClean from 'd2-utilizr/lib/arrayClean';
 import { itemTypeMap } from '../itemTypes';
+import {
+    getListItemFields,
+    getFavoritesFields,
+    getIdNameFields,
+} from './metaData';
 
 // Helper functions
+
 export const onError = error => console.log('Error: ', error);
 
 export const getEndPointName = type => itemTypeMap[type].endPointName;
 
-// Fields
-export const getIdNameFields = ({ rename } = {}) => [
-    'id',
-    `${rename ? 'displayName~rename(name)' : 'name,displayName'}`,
-];
-
-export const getItemFields = () => ['dimensionItem~rename(id)'];
-
-export const getDimensionFields = ({ withItems }) =>
-    arrayClean([
-        'dimension',
-        'legendSet[id]',
-        withItems ? `items[${getItemFields().join(',')}]` : ``,
-    ]);
-
-export const getAxesFields = ({ withItems }) => [
-    `columns[${getDimensionFields({ withItems }).join(',')}]`,
-    `rows[${getDimensionFields({ withItems }).join(',')}]`,
-    `filters[${getDimensionFields({ withItems }).join(',')}]`,
-];
-
-export const interpretationFields = () => {
-    return 'interpretations[id]';
-    // return 'interpretations[id,text,created,user[id,displayName],likedBy,comments[id,text,created,user[id,displayName]]]';
-};
-
-export const getFavoriteFields = ({ withDimensions, withOptions }) => {
-    return arrayClean([
-        `${getIdNameFields({ rename: true }).join(',')}`,
-        'displayDescription~rename(description)',
-        interpretationFields(),
-        withDimensions ? `${getAxesFields({ withItems: true }).join(',')}` : ``,
-        withOptions
-            ? [
-                  '*',
-                  '!attributeDimensions',
-                  '!attributeValues',
-                  '!category',
-                  '!categoryDimensions',
-                  '!categoryOptionGroupSetDimensions',
-                  '!columnDimensions',
-                  '!dataDimensionItems',
-                  '!dataElementDimensions',
-                  '!dataElementGroupSetDimensions',
-                  '!filterDimensions',
-                  '!itemOrganisationUnitGroups',
-                  '!lastUpdatedBy',
-                  '!organisationUnitGroupSetDimensions',
-                  '!organisationUnitLevels',
-                  '!organisationUnits',
-                  '!programIndicatorDimensions',
-                  '!relativePeriods',
-                  '!reportParams',
-                  '!rowDimensions',
-                  '!series',
-                  '!translations',
-                  '!userOrganisationUnit',
-                  '!userOrganisationUnitChildren',
-                  '!userOrganisationUnitGrandChildren',
-              ].join(',')
-            : '',
-    ]);
-};
-
-export const getFavoritesFields = ({ withDimensions, withOptions }) => [
-    `reportTable[${getFavoriteFields({ withDimensions }).join(',')}]`,
-    `chart[${getFavoriteFields({ withDimensions }).join(',')}]`,
-    `map[${getFavoriteFields({ withDimensions }).join(',')}]`,
-    `eventReport[${getFavoriteFields({ withDimensions }).join(',')}]`,
-    `eventChart[${getFavoriteFields({ withDimensions }).join(',')}]`,
-];
-
-export const getListItemFields = () => [
-    `reports[${getIdNameFields({ rename: true }).join(',')}]`,
-    `resources[${getIdNameFields({ rename: true }).join(',')}]`,
-    `users[${getIdNameFields({ rename: true }).join(',')}]`,
-];
-
+// Dashboard item
 export const getDashboardItemsFields = ({ withFavorite } = {}) =>
     arrayClean([
         'id',
@@ -104,6 +33,7 @@ export const getDashboardItemsFields = ({ withFavorite } = {}) =>
             : ``,
     ]);
 
+// Dashboard
 export const getDashboardFields = ({ withItems, withFavorite } = {}) =>
     arrayClean([
         `${getIdNameFields().join(',')}`,
